Refactor ledger tree structs into the ledger folder
