+++ dpc/src/ledger/record_proof.rs
