+++ dpc/src/record/mod.rs
@@ -26,8 +26,5 @@ pub use payload::*;
 pub mod record;
 pub use record::*;
 
-pub mod record_proof;
-pub use record_proof::*;
-
 #[cfg(test)]
 mod tests;
